move getTLSConfig back to util
