+++ internal/controller/utils.go
@@ -17,12 +17,17 @@ limitations under the License.
 package controller
 
 import (
+	"context"
+	"crypto/tls"
+	"crypto/x509"
 	"fmt"
 	"maps"
 	"slices"
 	"strconv"
 
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	valkeyv1 "valkey.io/valkey-operator/api/v1alpha1"
 	"valkey.io/valkey-operator/internal/valkey"
 )
@@ -236,4 +241,31 @@ func countSlots(ranges []valkey.SlotsRange) int {
 // replicas.
 func valkeyNodeName(clusterName string, shardIndex int, nodeIndex int) string {
 	return fmt.Sprintf("%s-%d-%d", clusterName, shardIndex, nodeIndex)
+}
+
+// getTLSConfig returns the TLS configuration for a ValkeyCluster.
+func getTLSConfig(ctx context.Context, c client.Client, secretName, serverName, namespace string) (*tls.Config, error) {
+	secret := &corev1.Secret{}
+	err := c.Get(ctx, client.ObjectKey{Namespace: namespace, Name: secretName}, secret)
+	if err != nil {
+		return nil, err
+	}
+
+	caData, caOk := secret.Data[tlsSecretKeyCA]
+
+	if !caOk {
+		return nil, fmt.Errorf("TLS secret is missing required key: ca=%v", caOk)
+	}
+
+	caCertPool := x509.NewCertPool()
+	if !caCertPool.AppendCertsFromPEM(caData) {
+		return nil, fmt.Errorf("failed to parse CA certificates from secret key %q", "ca.crt")
+	}
+
+	tlsCfg := &tls.Config{
+		RootCAs:    caCertPool,
+		ServerName: serverName,
+		MinVersion: tls.VersionTLS12,
+	}
+	return tlsCfg, nil
 }
